Still bugs

+++ GraphColoring/src/space/SpaceImpl.java
@@ -108,7 +108,7 @@ public class SpaceImpl extends UnicastRemoteObject implements Space {
             return;
         }
         if (!closures.containsKey(closure.getId())){
-            if (!closure.isAncestor() && closure.getHeuristic() >= (Integer)global.getValue()){
+            if (!closure.isAncestor() && closure.getHeuristic() >= (Double)global.getValue()){
                 putResult(new Result(((Continuation)closure.getArgument(0)).argument, 0));
             }
             closures.put(closure.getId(), closure);
